Add support for GGML file loading.
This is a (very) direct port of the llama_file_loader and llama_model_loader functions from llama.cpp.
This is a port of the older code which was a bit simpler.  There was a rewrite that happened right around
the time I started this work, so at some point this will have to be redone, but for the moment this should
be compatible with the vast majority of ggml models on the internet.

I began the work to refactor the load system to be asynchronous, but as you can see it is very half-assed
and needs a lot more work.