Break up matmul into smaller pieces for massive performance gains!
It was convenient to make matmul use one thread per line, but that isn't super efficient from either a thread density or memory access perspective.  We now break up matmul into threadgroups of 32x32 where each thread processes 32 values.  We still cache the input vector in groupshared memory, but only the pieces needed for that particular group.

The end result is that matmul time went from 70ms to 20ms on the 3b model!  This is down from 80ms before the last change, so a really big speedup.  There is a bit more to be done with shader model 6 intrinsics, but actually all the little stuff is now taking most of the time, so the next step will be fusing some of the smaller kernels and fixing some of the most egregious performance issues.